requirements.txt
requests >= 2.31.0
beautifulsoup4 >= 4.12.3
pandas >= 2.2.1
numpy >= 1.26.4
Pillow >= 10.2.0
matplotlib >= 3.8.3
scikit-learn >= 1.4.1
umap-learn >= 0.5.5
wordcloud >= 1.9.3
opencv-python >= 4.9.0.80
ipykernel >= 6.29.3
ipympl >= 0.9.3
lxml >= 5.1.0
pyarrow >= 15.0.0
fastparquet >= 2024.2.0
pickleshare >= 0.7.5
torch >= 2.2.1
torchvision >= 0.17.1
ramp-workflow >= 0.6.0